backend/requirements.txt
flask==1.1.0
google-cloud-storage==1.35.0
gunicorn==20.0.4
pandas==1.1.5
traffic==2.5